Restructure tests for StringSeqCaseOps
